requirements.txt
openai==0.27.4
Pillow==9.5.0